MIFID-16

+++ main/src/main/java/pl/rg/window/users/UserWindowModel.java
@@ -3,13 +3,18 @@ package pl.rg.window.users;
 import java.awt.event.ActionListener;
 import java.util.ArrayList;
 import java.util.List;
+import javax.swing.JFrame;
+import javax.swing.JOptionPane;
 import javax.swing.JTable;
 import javax.swing.table.DefaultTableModel;
+import lombok.Getter;
+import pl.rg.users.UserDto;
 import pl.rg.users.UserModuleController;
 import pl.rg.utils.exception.ApplicationException;
 import pl.rg.utils.logger.LogLevel;
 import pl.rg.window.AbstractWindow;
 
+@Getter
 public class UserWindowModel extends AbstractWindow {
 
   private static final String[] buttonNames = {"Dodaj", "Edytuj", "Usuń", "Szukaj",
@@ -26,19 +31,26 @@ public class UserWindowModel extends AbstractWindow {
     this.userModuleController = userModuleController;
   }
 
-  public DefaultTableModel loadUserData() {
+  public DefaultTableModel loadUserData() { // todo replace after implement search users method
     DefaultTableModel tableModel = new DefaultTableModel(getColumnNames(), 0);
     tableModel.setRowCount(0);
-    Object[][] userData = { // todo replace after implement search users method
-        {1, "jankow", "Jan", "Kowalski", "jan.kowalski@example.com"},
-        {2, "annnow", "Anna", "Nowak", "anna.nowak@example.com"},
-        {3, "piowis", "Piotr", "Wiśniewski", "piotr.wisniewski@example.com"}
-    };
-    for (Object[] row : userData) {
-      tableModel.addRow(row);
+    List<UserDto> userData = userModuleController.getFiltered(null);
+    for (UserDto user : userData) {
+      tableModel.addRow(new Object[]{
+          user.getId(),
+          user.getUserName(),
+          user.getFirstName(),
+          user.getLastName(),
+          user.getEmail()
+      });
     }
 
     return tableModel;
+  }
+
+  public void refreshTable() {
+    DefaultTableModel userData = loadUserData();
+    mainTable.setModel(userData);
   }
 
   @Override
@@ -51,18 +63,30 @@ public class UserWindowModel extends AbstractWindow {
     actions = new ArrayList<>();
 
     ActionListener addAction = e -> {
-      UserWindow userWindow = new UserWindow(this);
+      UserWindow userWindow = new UserWindow(this, false, null);
       userWindow.setVisible(true);
     };
 
     ActionListener editAction = e -> {
-      // todo
+      int selectedRow = mainTable.getSelectedRow();
+      if (selectedRow < 0) {
+        JOptionPane.showMessageDialog(new JFrame(), "Nie wybrano żadnego użytkownika do edycji");
+        return;
+      }
+      int id = (int) mainTable.getValueAt(selectedRow, 0);
+      UserWindow userWindow = new UserWindow(this, true, id);
+      userWindow.setVisible(true);
     };
 
     ActionListener deleteAction = e -> {
       int selectedRow = mainTable.getSelectedRow();
+      if (selectedRow < 0) {
+        JOptionPane.showMessageDialog(new JFrame(), "Nie wybrano żadnego użytkownika do edycji");
+        return;
+      }
       Integer id = (Integer) mainTable.getValueAt(selectedRow, 0);
       userModuleController.deleteUser(id);
+      refreshTable();
     };
 
     ActionListener searchAction = e -> {
@@ -83,7 +107,6 @@ public class UserWindowModel extends AbstractWindow {
       throw logger.logAndThrowRuntimeException(LogLevel.DEBUG,
           new ApplicationException("M23SD", "Brak implementacji wszystkich akcji!"));
     }
-
   }
 
   @Override
